requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.31.1
pydantic>=2.5.0
boto3==1.34.0
python-dotenv==1.0.0
supabase>=2.9.0
requests==2.31.0
strands-agents==1.29.0
